requirements.txt
Django==4.2
psycopg==3.1.8
djangorestframework==3.14.0
pyyaml==6.0
uritemplate==4.1.1
django-rest-registration==0.8.0
flake8==6.0.0
isort==5.12.0
black==23.3.0
